Updated TextAPI SDK

+++ TextMood.Shared/Services/EmojiServices.cs
@@ -2,15 +2,15 @@
 {
     public static class EmojiServices
     {
-        public static string GetEmoji(float? sentimentScore)
+        public static string GetEmoji(double? sentimentScore)
         {
             switch (sentimentScore)
             {
-                case float number when (number >= 0 && number < 0.4):
+                case double number when (number >= 0 && number < 0.4):
                     return EmojiConstants.SadFaceEmoji;
-                case float number when (number >= 0.4 && number <= 0.6):
+                case double number when (number >= 0.4 && number <= 0.6):
                     return EmojiConstants.NeutralFaceEmoji;
-                case float number when (number > 0.6):
+                case double number when (number > 0.6):
                     return EmojiConstants.HappyFaceEmoji;
                 case null:
                     return EmojiConstants.BlankFaceEmoji;
